EfficientNet/requirements.txt
h5py
joblib
numpy
pandas
Pillow
scikit_learn
tensorflow
tensorflow_intel
torch
ultralytics
